Update web-api.md
errorHandler() what returning an Object in one of the examples, which is misleading because in Spark ExceptionHandlers should return void. If errorHandler() does return anything, it is not used. 

I spent hours trying to figure out why my exception handlers weren't setting a response body, and this example code was what sent me down the wrong direction because I thought the response body would be set to whatever my exception handlers returned, which is not the case.